pyproject.toml
[tool.poetry]
name = "woo-sy"
version = "0.1.2"
description = "WooCommerce + Etsy = woo-sy | A tool for syncing listings between the two"
authors = ["Alexander VanTol <alexander.m.vantol@gmail.com>"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "^3"
woocommerce = "^1.2"
etsy_py = "^1.0"
python-wordpress-xmlrpc = "^2.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.black]
line-length = 88
py36 = true
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | dist
)/
'''